refactor(data)!: unify persisted literals and runtime scalar values
Merge tabular contracts into yss-data-contract and share checked scalar carriers across kernels, relational execution and results. Use one literal tree for graph constants and protocol defaults, preserve wide integers on the wire, and update frontend editors and parsers together.

BREAKING CHANGE: persisted and transport data values use the unified tagged literal representation; table constants retain snapshots without derived resource handles.

+++ src-tauri/crates/yss-database-arrow/src/conversion.rs
@@ -12,12 +12,12 @@ use std::{
     hash::{Hash, Hasher},
     sync::Arc,
 };
+use yss_data_contract::TabularScalar;
 use yss_data_contract::{
     ColumnSemantic, ConversionMetadata, DatetimeRepresentation, NumericConstraints,
     NumericRepresentation, SemanticConversion, SemanticType, SemanticValue, TemporalPrecision,
     TemporalType,
 };
-use yss_tabular_contract::TabularScalar;
 
 pub struct ConvertedValues {
     pub values: Vec<TabularScalar>,
@@ -143,7 +143,7 @@ fn materialized_values(array: &dyn Array) -> Result<Vec<TabularScalar>, TabularA
             .iter()
             .map(|value| match value {
                 None => Ok(TabularScalar::Null),
-                Some(value) => Ok(TabularScalar::Decimal(
+                Some(value) => Ok(TabularScalar::Float64(
                     value
                         .try_into()
                         .map_err(|_| TabularArrowError::InvalidValue)?,
